Collapse climb/elevetor classes into lifter
After doing some mashing on climb and elevator classes, I decided that
there was just oo much common code that was being duplicated and could
be better handled in a single class or some other abstraction.  I still
don't think the partitioning is correct but better than it was.  I
haven't deleted the other classes as we may revert but for now we will
go with this design