Fixing an error that slipped through
There was an error that was let through somehow (it was just one character missing, and I fixed it earlier in the day, but who knows), which is why I hate uploading to github. This should wrap up the latest code form the competition today, I just checked it and it's all good.